feat(users): refactor users app
