+++ taskforce/apps/users/src/app/auth/auth.module.ts
@@ -3,9 +3,10 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { getJwtConfig } from '../../config/jwt.config';
+import { UserModule } from '../user/user.module';
+import { UserService } from '../user/user.service';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import {UserModule} from "../user/user.module";
 import { JwtStrategy } from './strategies/jwt.strategy';
 
 @Module({
@@ -19,6 +20,7 @@ import { JwtStrategy } from './strategies/jwt.strategy';
       })
     ],
   controllers: [AuthController],
-  providers: [AuthService, JwtStrategy],
+  exports: [AuthService],
+  providers: [AuthService, UserService, JwtStrategy],
 })
 export class AuthModule {}
