+++ taskforce/apps/users/src/app/auth/dto/login-user.dto.ts
@@ -1,28 +1,10 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsEmail, Length } from 'class-validator';
-import { AuthUserError, UserApiDescription, UserPasswordLength } from '../auth.constant';
+import { PickType } from '@nestjs/swagger';
+import CreateUserDto from '../../user/dto/create-user.dto';
 
-export class LoginUserDto {
-  @ApiProperty({
-    description: UserApiDescription.Email,
-    example: 'user@user.local'
-  })
-  @IsEmail(
-    {},
-    {
-      message: AuthUserError.EmailNotValid
-    })
-  public email: string;
-
-  @ApiProperty({
-    description: `User password, min ${UserPasswordLength.Min}, max ${UserPasswordLength.Max} chars length`,
-    example: '123456'
-  })
-  @Length(
-    UserPasswordLength.Min,
-    UserPasswordLength.Max,
-    {
-      message: AuthUserError.PasswordIsWrong
-    })
-  public password: string;
+export class LoginUserDto extends PickType(CreateUserDto, [
+  'email', 'password'
+] as const) {
+  constructor() {
+    super();
+  }
 }
