+++ taskforce/apps/users/src/app/user/dto/update-user.dto.ts
@@ -9,7 +9,7 @@ import {
   UserInfoLength,
   UserNameLength,
   UserOccupationCount,
-} from '../auth.constant';
+} from '../user.constant';
 
 export default class UpdateUserDto {
   @ApiProperty({
