+++ taskforce/apps/users/src/app/user/user.http
@@ -1,0 +1,51 @@
+# Create a new user
+POST http://localhost:3333/api/user/register HTTP/1.1
+Content-Type: application/json
+
+{
+  "name": "Keks",
+  "email": "user8@notfound.local",
+  "city": "Москва",
+  "password": "123456",
+  "dateBirth": "2003-12-22",
+  "role": "client"
+}
+###
+
+# Update user's password
+PATCH http://localhost:3333/api/user/password HTTP/1.1
+Content-Type: application/json
+
+{
+  "email": "user@notfound.local",
+  "currentPassword": "123456",
+  "newPassword": "234567"
+}
+###
+
+# Update user's avatar
+PUT http://localhost:3333/api/user/avatar HTTP/1.1
+Content-Type: application/json
+
+{
+  "email": "user@notfound.local",
+  "avatar": "/images/user.png"
+}
+###
+
+# Update user's data
+PUT http://localhost:3333/api/user/d04eb35d-c36f-4e2b-b828-136379c7c6e3 HTTP/1.1
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2Mzk3MTNhMmIzYjk2ZDA5OTg0OWQ5MjMiLCJlbWFpbCI6InVzZXJAbm90Zm91bmQubG9jYWwiLCJyb2xlIjoiY2xpZW50IiwibmFtZSI6Iktla3MiLCJpYXQiOjE2NzExNzgzNTIsImV4cCI6MTY3MTE3ODQxMn0.GeftGefxDrETnaG2zCOHI8CDp-DPzDaMMC3Crf7noys
+Content-Type: application/json
+
+{
+  "name": "Felix",
+  "city": "Санкт-Петербург",
+  "dateBirth": "2014-04-22"
+}
+###
+
+# Get user by ID
+GET http://localhost:3333/api/user/d04eb35d-c36f-4e2b-b828-136379c7c6e3 HTTP/1.1
+
+###
