+++ taskforce/apps/users/src/app/validators/age-validator.ts
@@ -1,6 +1,6 @@
 import { ValidatorConstraint, ValidatorConstraintInterface } from 'class-validator';
 import * as dayjs from 'dayjs';
-import { UserAge } from '../auth/auth.constant';
+import { UserAge } from '../user/user.constant';
 
 @ValidatorConstraint({ name: 'Invalid Age', async: false })
 export class AgeValidator implements ValidatorConstraintInterface {
