+++ taskforce/environments/.users.env-example
@@ -4,3 +4,4 @@ MONGO_PORT=27017
 MONGO_USER=admin
 MONGO_PASSWORD=test
 MONGO_AUTH_BASE=admin
+JWT_SECRET=secret
