+++ taskforce/package.json
@@ -18,17 +18,17 @@
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/swagger": "6.1.3",
     "@prisma/client": "4.6.1",
-    "reflect-metadata": "^0.1.13",
     "bcrypt": "5.1.0",
-    "rxjs": "^7.0.0",
-    "tslib": "^2.3.0",
     "class-transformer": "^0.5.1",
     "class-validator": "0.13.2",
+    "dayjs": "1.11.6",
     "mongoose": "6.7.2",
     "passport": "0.6.0",
     "passport-jwt": "4.0.0",
     "passport-local": "1.0.0",
-    "dayjs": "1.11.6"
+    "reflect-metadata": "^0.1.13",
+    "rxjs": "^7.0.0",
+    "tslib": "^2.3.0"
   },
   "devDependencies": {
     "@nestjs/schematics": "^9.0.0",
@@ -42,6 +42,7 @@
     "@nrwl/workspace": "15.1.0",
     "@types/bcrypt": "5.0.0",
     "@types/jest": "28.1.1",
+    "@types/multer": "^1.4.7",
     "@types/node": "18.7.1",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
